requirements.txt
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0
aiohttp>=3.8.0
rank-bm25>=0.2.2
nltk>=3.8.1